Update conversao_metros_cm.py
Update file

+++ conversao_metros_cm.py
@@ -1,5 +1,5 @@
 #Deve-se inserir um valor numérico abaixo
-metros = int(input('Informe o valor em metros: '))
+metros = float(input('Informe o valor em metros: '))
 
 
 #Temos a resposta do valor já convertido para centímetros
